refactor: fix naming to be more verbose and readable
