+++ src/app/redux/search-item-reducer/search-item.actions.ts
@@ -1,24 +1,24 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { createAction, props } from '@ngrx/store';
-import { ResponseVideo } from 'src/app/youtube/models/yt-models';
+import { ResponseVideo } from 'src/app/youtube/models/you-tube-models';
 import { SortType } from './search-item.reducer';
 
 export const updateSearchItems = createAction(
-  '[Search Results Page] Yt Search Success',
+  '[Search Results Page] Youtube Search Success',
   props<{ searchItems: ResponseVideo[] }>()
 );
 
 export const searchItemsError = createAction(
-  '[Search Results Page] Yt Search Error',
+  '[Search Results Page] Youtube Search Error',
   props<{ error: HttpErrorResponse }>()
 );
 
 export const clearSearchItemsError = createAction(
-  '[Search Results Page] clearing Yt Search Error '
+  '[Search Results Page] clearing Youtube Search Error '
 );
 
 export const initiateSearch = createAction(
-  '[Search Results Page] Initiate Yt Search ',
+  '[Search Results Page] Initiate Youtube Search ',
   props<{ searchQuery: string }>()
 );
