+++ src/app/redux/user-items-reducer/user-items.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { UserCardInfo } from 'src/app/youtube/models/yt-models';
+import { UserCardInfo } from 'src/app/youtube/models/you-tube-models';
 
 export const updateUserItems = createAction(
   '[User Items Page] Card creation',
