+++ src/app/redux/user-items-reducer/user-items.reducer.ts
@@ -1,5 +1,5 @@
 import { createReducer, on } from '@ngrx/store';
-import { UserCardInfo } from 'src/app/youtube/models/yt-models';
+import { UserCardInfo } from 'src/app/youtube/models/you-tube-models';
 import { updateUserItems } from './user-items.actions';
 
 export const initialState: UserCardInfo[] | [] = [];
