+++ src/app/redux/user-items-reducer/user-items.selector.ts
@@ -1,5 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { UserCardInfo } from 'src/app/youtube/models/yt-models';
+import { UserCardInfo } from 'src/app/youtube/models/you-tube-models';
 
 const selectUserItemsFeature =
   createFeatureSelector<UserCardInfo[]>('userItems');
