+++ src/app/youtube/components/user-item/user-item.component.html
@@ -1,8 +1,8 @@
 <img class="card-s-img" src="{{imgUrl}}" />
 <div class="card-s-info">
-  <p>{{ userVid.date }}</p>
+  <p>{{ userVideo.date }}</p>
 </div>
-<h3 class="card-s-name">{{ userVid.title }}</h3>
-<a href="{{ userVid.link }}" target="_blank" class="card-s-more button-blue">
+<h3 class="card-s-name">{{ userVideo.title }}</h3>
+<a href="{{ userVideo.link }}" target="_blank" class="card-s-more button-blue">
   More
 </a>
