+++ src/app/youtube/pages/search-page/search-page.component.html
@@ -1,10 +1,10 @@
 <app-header></app-header>
 <div class="main">
   <div class="search-list">
-    <ng-container *ngIf="searchItemsFacade.searchItems$ | async as vids">
-      <app-search-item *ngFor="let vid of vids let index=index; trackBy:vidById" [vidObj]="vid">
+    <ng-container *ngIf="searchItemsFacade.searchItems$ | async as videos">
+      <app-search-item *ngFor="let video of videos let index=index; trackBy:videoById" [video]="video">
       </app-search-item>
-      <h2 *ngIf="vids.length === 0">
+      <h2 *ngIf="videos.length === 0">
         Nothing to display, try searching for something
       </h2>
     </ng-container>
