+++ src/app/youtube/pages/user-items-page/user-items-page.component.ts
@@ -1,7 +1,7 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { Observable } from 'rxjs';
 import { UserItemsFacadeService } from 'src/app/redux/user-items-reducer/user-items-facade.service';
-import { UserCardInfo } from '../../models/yt-models';
+import { UserCardInfo } from '../../models/you-tube-models';
 
 @Component({
   selector: 'app-user-items-page',
